\name{axisTrack}
\alias{axisTrack}
\alias{cpgIslands}
\alias{refGenes}
\docType{data}
\title{
  Example data for plotting annotation.
}
\description{
  Five annotation tracks for plotting in Gviz.
}
\usage{
  data(axisTrack)
  data(cpgIslands)
  data(refGenes)
}

\details{
  These tracks are based on genome="hg19", chr = "chr11", 
  start = 31000000L, end =  33000000L.
}
\examples{
data(axisTrack)
data(cpgIslands)
data(refGenes)
}
\keyword{datasets}
