\name{axtDanRer7Hg19}
\alias{axtDanRer7Hg19}
\alias{axtHg19DanRer7}
\docType{data}
\title{
The dataset axtDanRer7Hg19, axtHg19DanRer7
}
\description{
  The example CNEs from part of hg19 and danRer7 comparison.
}

\usage{data(axtDanRer7Hg19)
       data(axtHg19DanRer7)}


\examples{
  data(axtDanRer7Hg19)
}
\keyword{datasets}
