\name{cneBlatedDanRer7Hg19}
\alias{cneBlatedDanRer7Hg19}
\docType{data}
\title{
The dataset cneBlatedDanRer7Hg19
}
\description{
This example dataset is the CNEs between hg19 and danRer7 after 
running blat program 
at the thresholds "45_50", "48_50" and "49_50".
}

\usage{data(cneBlatedDanRer7Hg19)}


\examples{
data(cneBlatedDanRer7Hg19)

}
\keyword{datasets}
