\name{CNEDanRer10Hg38}
\alias{CNEDanRer10Hg38}
\alias{CNEHg38DanRer10}
\docType{data}
\title{
CNEHg38DanRer10 and CNEDanRer10Hg38 dataset
}
\description{
These two datasets are the direct output from \code{ceScan}.
}
\usage{data(CNEHg38DanRer10)}

\examples{
data(CNEHg38DanRer10)
}
\keyword{datasets}
