\name{cneFinalListDanRer10Hg38}
\alias{cneFinalListDanRer10Hg38}
\docType{data}
\title{
  cneFinalListDanRer10Hg38 dataset
}
\description{
  cneFinalListDanRer10Hg38 dataset contains the CNE between danRer10 and hg38
  around chr6:24,000,000..27,000,000.
}
\usage{data("cneFinalListDanRer10Hg38")}

\examples{
  data(cneFinalListDanRer10Hg38)
}
