\name{cneHg19DanRer7_45_50}
\alias{cneHg19DanRer7_45_50}
\alias{cneHg19DanRer7_48_50}
\alias{cneHg19DanRer7_49_50}
\docType{data}
\title{
  These datasets of CNE density values.
}
\description{
  These three datasets are output from \code{CNEDensity}.
}
\usage{data(cneHg19DanRer7_45_50)}

\examples{
data(cneHg19DanRer7_45_50)
data(cneHg19DanRer7_48_50)
data(cneHg19DanRer7_49_50)
}
\keyword{datasets}
