\name{finalCNE}
\alias{finalCNE}
\docType{data}
\title{
 finalCNE dataset
}
\description{
 One example dataset in CNE class.
}
\usage{data(finalCNE)}

\details{
  This is a subset of CNEs between hg19 and danRer7 on chromosome 11,
  from 31000000L to 32500000L based on hg19 coordinate.
}
\examples{
data(finalCNE)
}
\keyword{datasets}
