\name{grangesPairsForDotplot}
\alias{grangesPairsForDotplot}
\docType{data}
\title{
  grangesPairsForDotplot
}
\description{
  Example of \code{GrangePairs} object from 
  the collineary regions of \emph{Adineta vaga}.
}
\usage{data("grangesPairsForDotplot")}

\details{
  The collineary regions from \dQuote{scaffold_1} and \dQuote{scaffold_5}.
}

\source{
  Example from own project.
}
\examples{
  data(grangesPairsForDotplot)
}
\keyword{datasets}