\name{qSizesDanRer7}
\alias{qSizesDanRer7}
\alias{qSizesHg19}
\docType{data}
\title{
  The chromosome sizes data.
}
\description{
  The chromosome sizes data of hg19 and danRer7.
}

\usage{data(qSizesDanRer7)
       data(qSizesHg19)
}

\source{
  http://hgdownload.soe.ucsc.edu/downloads.html
}
\examples{
data(qSizesDanRer7)
data(qSizesHg19)
}
\keyword{datasets}
