\name{axisTrack}
\alias{axisTrack}
\alias{cpgIslands}
\alias{refGenes}
\docType{data}
\title{
  Example data for plotting annotation.
}
\description{
  Five annotation tracks for plotting in Gviz.
}
\usage{
  data(axisTrack)
  data(cpgIslands)
  data(refGenes)
}

\details{
  These tracks are based on genome="danRer10", chr = "chr6", 
  start = 24000000, end = 27000000. 
}
\examples{
data(axisTrack)
data(cpgIslands)
data(refGenes)
}
\keyword{datasets}
